Rename weaviate.py -> weaviate_client.py
