+++ dlt/destinations/weaviate/__init__.py
@@ -44,7 +44,7 @@ def capabilities() -> DestinationCapabilitiesContext:
 def client(
     schema: Schema, initial_config: DestinationClientConfiguration = config.value
 ) -> JobClientBase:
-    from dlt.destinations.weaviate.weaviate import WeaviateClient
+    from dlt.destinations.weaviate.weaviate_client import WeaviateClient
 
     return WeaviateClient(schema, _configure(initial_config))  # type: ignore
